requirements/prd.txt
-r base.txt

dj-database-url==0.3.0
django-redis==4.5.0
raven==5.29.0
uwsgi==2.0.11.2
